printout fun spheres

+++ src/main/scala/org/bbstilson/raytracer/models/Scale.scala
@@ -6,7 +6,7 @@ import MatrixDouble._
 class Scale(m: Matrix) extends MatrixDouble(m)
 
 object Scale {
-  def apply(x: Int, y: Int, z: Int): Scale = {
+  def apply(x: Double, y: Double, z: Double): Scale = {
     val vals = Vector(x, y, z, 1)
     val m = mkMatrix(4, 4, (r, c) => if (r == c) vals(r) else 0d)
     new Scale(m)
